refactor dependency directon
